Printf now handling almost like the real one

+++ Ft_printf/src/print_arg_char.c
@@ -6,7 +6,7 @@
 /*   By: wwatkins <wwatkins@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/02/19 17:52:21 by wwatkins          #+#    #+#             */
-/*   Updated: 2016/03/03 18:09:36 by wwatkins         ###   ########.fr       */
+/*   Updated: 2016/03/04 12:15:20 by wwatkins         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ void	print_arg_char(t_e *e, t_a *arg, int *i)
 		e->plen++;
 	}
 	if (arg->mod.l || arg->type == 'C')
-		*i = print_wchar_t(c);
+		*i = print_wchar_t((unsigned int)c);
 	else
 		*i = print_char(va_arg(e->ap, int));
 	while (arg->flag.mn && arg->width-- > 0)
